# internal/adapters/http/api_types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package api

import (
	"time"

	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// CreateAccountRequest defines model for CreateAccountRequest.
type CreateAccountRequest struct {
	Currency Currency `json:"currency"`
}

// Currency defines model for Currency.
type Currency = interface{}

// Customer defines model for Customer.
type Customer struct {
	CreatedAt *time.Time          `json:"created_at,omitempty"`
	Email     *string             `json:"email,omitempty"`
	Uuid      *openapi_types.UUID `json:"uuid,omitempty"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// GetAccountResponse defines model for GetAccountResponse.
type GetAccountResponse struct {
	AccountNumber *string             `json:"account_number,omitempty"`
	Balance       *float64            `json:"balance,omitempty"`
	Currency      *Currency           `json:"currency,omitempty"`
	CustomerId    *openapi_types.UUID `json:"customer_id,omitempty"`
}

// RefreshTokenRequest defines model for RefreshTokenRequest.
type RefreshTokenRequest struct {
	RefreshToken string `json:"refresh_token"`
}

// ReplenishRequest defines model for ReplenishRequest.
type ReplenishRequest struct {
	Amount   float64  `json:"amount"`
	Currency Currency `json:"currency"`
	FromCard string   `json:"from_card"`
}

// SignInRequest defines model for SignInRequest.
type SignInRequest struct {
	Email string `json:"email"`

	// Password A password should be at least 8 characters long only with numbers and letters
	Password string `json:"password"`
}

// SignInResponse defines model for SignInResponse.
type SignInResponse struct {
	IdToken      *string `json:"id_token,omitempty"`
	RefreshToken *string `json:"refresh_token,omitempty"`
}

// SignUpRequest defines model for SignUpRequest.
type SignUpRequest struct {
	Email string `json:"email"`

	// Password A password should be at least 8 characters long only with numbers and letters
	Password string `json:"password"`
}

// TransferRequest defines model for TransferRequest.
type TransferRequest struct {
	Amount    float64 `json:"amount"`
	ToAccount string  `json:"to_account"`
}

// WithdrawRequest defines model for WithdrawRequest.
type WithdrawRequest struct {
	Amount float64 `json:"amount"`
	ToCard string  `json:"to_card"`
}

// AccountCreateJSONRequestBody defines body for AccountCreate for application/json ContentType.
type AccountCreateJSONRequestBody = CreateAccountRequest

// AccountReplenishJSONRequestBody defines body for AccountReplenish for application/json ContentType.
type AccountReplenishJSONRequestBody = ReplenishRequest

// AccountTransferJSONRequestBody defines body for AccountTransfer for application/json ContentType.
type AccountTransferJSONRequestBody = TransferRequest

// AccountWithdrawJSONRequestBody defines body for AccountWithdraw for application/json ContentType.
type AccountWithdrawJSONRequestBody = WithdrawRequest

// RefreshTokenJSONRequestBody defines body for RefreshToken for application/json ContentType.
type RefreshTokenJSONRequestBody = RefreshTokenRequest

// CustomerSignInJSONRequestBody defines body for CustomerSignIn for application/json ContentType.
type CustomerSignInJSONRequestBody = SignInRequest

// CustomerSignUpJSONRequestBody defines body for CustomerSignUp for application/json ContentType.
type CustomerSignUpJSONRequestBody = SignUpRequest